Fixed nextjs hydration mismatch error
My layout component included a body tag (because this is what I've done in other frameworks in the past).
NextJs wraps your whole app in body>div with id __next
So the DOM is body>div>body which is semantically wrong but more importantly for the error in question you can't nest a body inside a div. The browser console even tells me this.
Oops.